activity_main.xml
Trying to get this set up.

I think that tomorrow I want to get it looking right, and then sometime soon figure out how to implement firestore into it.

I think that this might be pretty easy to get looking good, I just have to figure out how to put icons in for buttons, or figure out onClicks for icons, and then do gradients and stuff like that.